+++ src/actions/WalletActions.js
@@ -26,6 +26,14 @@ export type WalletAction =
           pathname?: string,
       }
     | {
+          type: typeof WALLET.SET_HIDDEN_COINS,
+          hiddenCoins: Array<string>,
+      }
+    | {
+          type: typeof WALLET.SET_HIDDEN_COINS_EXTERNAL,
+          hiddenCoinsExternal: Array<string>,
+      }
+    | {
           type: typeof WALLET.TOGGLE_DEVICE_DROPDOWN,
           opened: boolean,
       }