+++ src/views/Wallet/components/LeftNavigation/components/CoinMenu/index.messages.js
@@ -7,6 +7,15 @@ const definedMessages: Messages = defineMessages({
         id: 'TR_OTHER_COINS',
         defaultMessage: 'Other coins',
     },
+    TR_SELECT_COINS: {
+        id: 'TR_SELECT_COINS',
+        description: 'COMPLETE SENTENCE: Select a coin in application settings',
+        defaultMessage: 'Select a coin in {TR_SELECT_COINS_LINK}',
+    },
+    TR_SELECT_COINS_LINK: {
+        id: 'TR_SELECT_COINS_LINK',
+        defaultMessage: 'application settings',
+    },
 });
 
 export default definedMessages;